#2 - move: filteredMessagesData
