+++ src/index.tsx
@@ -19,7 +19,7 @@ import MessageStructure from './model/Message';
 import { PersonStructure } from './model/Person.ts';
 import dayjs from 'dayjs';
 import { getRandomPerson } from './helpers/getRandomPerson.ts';
-import { SearchProvider, useSearch } from './context/SearchContext.tsx';
+import { SearchProvider } from './context/SearchContext.tsx';
 
 const Main = () => {
   const [messagesData, setMessagesData] = useState<MessageStructure[]>([]);
@@ -31,7 +31,6 @@ const Main = () => {
   });
 
   const now = dayjs().format('YYYY-MM-DD HH:mm:ss');
-  const { searchQuery } = useSearch();
 
   useEffect(() => {
     const storedMessagesData = sessionStorage.getItem('messagesData');
@@ -157,16 +156,6 @@ const Main = () => {
     });
   };
 
-  const filterMessagesData = (
-    messagesData: MessageStructure[],
-    searchTerm: string
-  ) =>
-    messagesData.filter((message) =>
-      message.text.toLowerCase().includes(searchTerm.toLowerCase())
-    );
-
-  const filteredMessagesData = filterMessagesData(messagesData, searchQuery);
-
   const router = createBrowserRouter(
     createRoutesFromElements(
       <Route path="/" element={<App />}>
@@ -174,7 +163,7 @@ const Main = () => {
           path="/"
           element={
             <HomePage
-              messagesData={filteredMessagesData}
+              messagesData={messagesData}
               setMessagesData={setMessagesData}
               randomPerson={randomPerson}
               handleClickLike={handleClickLike}
